backend/requirements.txt
fastapi
psycopg2-binary
uvicorn
sqlalchemy
python-dotenv
starlette-session
bcrypt

# Dependecies for AI providers
openai
genai
requests
anthropic

# Dependencies for testing
pytest
httpx